Refactored Chat Messages fetching
Instead of being sent upon a joinRoom event in the backend, messages are sent in batch through a single getRoomMessagesPage call that is currently paired with a joinRoom emission in the frontend.

Instead, joinRoom now returns a ChatRoomEntity object, with the corresponding room info.

TODO: add a roomsInfo to the model, and switch room list over to that. Make sure to properly update the list when new rooms are joined or left